kafka_to_mongo/requirements.txt
kafka-python
pymongo
pandas